# ENH: Add analysis of single results