Closes #370, #1452. Related to #789 but without leaning into the requirement of having to use `asv`'s environment management.

This is still a WIP, clearly it needs to be folded into the `Command` pattern used throughout ASV (or pending a full switch to `click`).