@HaoZeke 2025-01-09
Looks a bit like this:

```bash
python ../asv/comparer_bench.py .asv/results/rgx1gen11/a0f29428-conda-py3.11.json .asv/results/rgx1gen11/a0f29428-conda-py3.11-numpy.json .asv/results/benchmarks.json
Hello, comparing .asv/results/rgx1gen11/a0f29428-conda-py3.11.json and .asv/results/rgx1gen11/a0f29428-conda-py3.11-numpy.json!

All benchmarks:

| Change   | Before     | After     |   Ratio | Benchmark (Parameter)   |
|----------|------------|-----------|---------|-------------------------|
| +        | 34.0±0.1μs | 94.8±30μs |    2.79 | b [e]                   |
```